requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
click==8.1.7
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
MarkupSafe==3.0.2
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.33.0
Werkzeug==3.1.3
psycopg2-binary==2.9.10
